Fix hash-n-degree-quads 6
The `issue` function had a major/stupid bug! I had it return the given bnode instead of
the issued bnode if it had already issued it. The reason this didn't come up earlier is
that they've been the same!

Also, changed where we initialized chosen-path in order to not trip up the
next-permutation? calculation.