Move issuer into hndq-state
This resolves the weird "or" statement I had before. I'm using a reduce to simulate a
while loop that mutates the issuer. I was passing it as a function argument before, but
the algorithm calls for updates within the function, not just at the function
boundaries, so we can update its reference within the reduce.